requirements.txt
lxml
pandas
snowflake-connector-python>=2.4.1
apache-airflow-providers-snowflake==3.1.0
snowflake-sqlalchemy>=1.1.0
apache-airflow-providers-amazon
requests
urllib3
openpyxl
apache-airflow-providers-dbt-cloud==2.1.0